Renamed YamlDotNet.Analyzers to YamlDotNet.Analyzers.StaticGenerator
