+++ YamlDotNet.Analyzers.StaticGenerator/ClassObject.cs
@@ -20,38 +20,32 @@
 // SOFTWARE.
 
 using System;
+using System.Collections.Generic;
 using Microsoft.CodeAnalysis;
 
-namespace YamlDotNet.Analyzers
+namespace YamlDotNet.Analyzers.StaticGenerator
 {
-    public abstract class File
+    public class ClassObject
     {
-        private readonly Action<string> _write;
-        private readonly Action _indent;
-        private readonly Action _unindent;
+        public List<IFieldSymbol> FieldSymbols { get; }
+        public string FullName { get; }
+        public string GuidSuffix { get; }
+        public INamedTypeSymbol ModuleSymbol { get; }
+        public List<IPropertySymbol> PropertySymbols { get; }
+        public string SanitizedClassName { get; }
+        public bool IsDictionary { get; }
+        public bool IsList { get; }
 
-        public File(Action<string> write, Action indent, Action unindent, GeneratorExecutionContext context)
+        public ClassObject(string sanitizedClassName, INamedTypeSymbol moduleSymbol, bool isDictionary = false, bool isList = false)
         {
-            _write = write;
-            _indent = indent;
-            _unindent = unindent;
+            FieldSymbols = new List<IFieldSymbol>();
+            PropertySymbols = new List<IPropertySymbol>();
+            FullName = moduleSymbol.GetFullName() ?? string.Empty;
+            GuidSuffix = Guid.NewGuid().ToString("N");
+            ModuleSymbol = moduleSymbol;
+            SanitizedClassName = sanitizedClassName;
+            IsDictionary = isDictionary;
+            IsList = isList;
         }
-
-        public void Write(string text)
-        {
-            _write(text);
-        }
-
-        public void Indent()
-        {
-            _indent();
-        }
-
-        public void UnIndent()
-        {
-            _unindent();
-        }
-
-        public abstract void Write(ClassSyntaxReceiver classSyntaxReceiver);
     }
 }
