+++ YamlDotNet.Analyzers.StaticGenerator/ClassSyntaxReceiver.cs
@@ -25,7 +25,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace YamlDotNet.Analyzers
+namespace YamlDotNet.Analyzers.StaticGenerator
 {
     public class ClassSyntaxReceiver : ISyntaxContextReceiver
     {
