+++ YamlDotNet.Analyzers.StaticGenerator/StaticPropertyDescriptorFile.cs
@@ -20,11 +20,9 @@
 // SOFTWARE.
 
 using System;
-using System.Collections.Generic;
-using System.Text;
 using Microsoft.CodeAnalysis;
 
-namespace YamlDotNet.Analyzers
+namespace YamlDotNet.Analyzers.StaticGenerator
 {
     public class StaticPropertyDescriptorFile : File
     {
